Split AdventureContext logic to custom hooks
